refactor save/load stuff into new sourmash.save_load module
